slightly optimized interface value accessor
